requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
selenium==4.15.0
openai==1.3.0
schedule==1.2.0
python-dotenv==1.0.0
lxml==4.9.3
pandas==2.1.3
dataclasses-json==0.6.1
aiohttp==3.9.0
webdriver-manager==4.0.1
pytz==2023.3
pyyaml==6.0.1
fake-useragent==1.4.0 